// src/api/generated/schemas/providerSettingDefinitionResponse.ts
/**
 * Generated by orval v7.12.2 🍺
 * Do not edit manually.
 * Notifiable API
 * A simple example ASP.NET Core Web API
 * OpenAPI spec version: v1
 */
import type { SettingFieldType } from './settingFieldType';

export interface ProviderSettingDefinitionResponse {
  key?: string;
  displayName?: string;
  /** @nullable */
  description?: string | null;
  fieldType?: SettingFieldType;
  isRequired?: boolean;
  /** @nullable */
  placeholder?: string | null;
  /** @nullable */
  defaultValue?: string | null;
}