// src/api/generated/schemas/settingFieldType.ts
/**
 * Generated by orval v7.12.2 🍺
 * Do not edit manually.
 * Notifiable API
 * A simple example ASP.NET Core Web API
 * OpenAPI spec version: v1
 */

export type SettingFieldType = (typeof SettingFieldType)[keyof typeof SettingFieldType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SettingFieldType = {
  Text: 'text',
  Secret: 'secret',
  Email: 'email',
  Url: 'url',
  Json: 'json',
  Number: 'number',
  Boolean: 'boolean',
} as const;